Refactored api module to common - event bus transfer objects under dataobjects package, handlebars extension under handlebars, launcher under launcher package
