+++ knotx-core/knotx-common/src/main/java/com/cognifide/knotx/dataobjects/TemplateEngineRequest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.cognifide.knotx.api;
+package com.cognifide.knotx.dataobjects;
 
 import com.google.common.cache.Cache;
 import com.google.common.cache.CacheBuilder;
