+++ knotx-core/knotx-repository/src/main/java/com/cognifide/knotx/repository/RepositoryVerticle.java
@@ -19,8 +19,8 @@ package com.cognifide.knotx.repository;
 
 import com.google.common.collect.Lists;
 
-import com.cognifide.knotx.api.RepositoryRequest;
-import com.cognifide.knotx.api.RepositoryResponse;
+import com.cognifide.knotx.dataobjects.RepositoryRequest;
+import com.cognifide.knotx.dataobjects.RepositoryResponse;
 import com.cognifide.knotx.repository.impl.NullRepository;
 
 import org.apache.commons.lang3.tuple.Pair;
