+++ knotx-core/knotx-repository/src/main/java/com/cognifide/knotx/repository/impl/NullRepository.java
@@ -17,8 +17,8 @@
  */
 package com.cognifide.knotx.repository.impl;
 
-import com.cognifide.knotx.api.RepositoryRequest;
-import com.cognifide.knotx.api.RepositoryResponse;
+import com.cognifide.knotx.dataobjects.RepositoryRequest;
+import com.cognifide.knotx.dataobjects.RepositoryResponse;
 import com.cognifide.knotx.repository.Repository;
 
 import io.netty.handler.codec.http.HttpResponseStatus;
