+++ knotx-core/knotx-standalone/pom.xml
@@ -34,7 +34,7 @@
 
   <properties>
     <!-- the main verticle class name -->
-    <main.verticle>com.cognifide.knotx.KnotxStarterVerticle</main.verticle>
+    <main.verticle>com.cognifide.knotx.launcher.KnotxStarterVerticle</main.verticle>
   </properties>
 
   <dependencies>
@@ -65,7 +65,7 @@
     <!-- Knot.x -->
     <dependency>
       <groupId>com.cognifide.knotx</groupId>
-      <artifactId>knotx-launcher</artifactId>
+      <artifactId>knotx-common</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
