+++ knotx-core/knotx-template-engine/src/main/java/com/cognifide/knotx/engine/TemplateEngineConsts.java
@@ -17,7 +17,7 @@
  */
 package com.cognifide.knotx.engine;
 
-public class TemplateEngineConsts {
+public interface TemplateEngineConsts {
   public static final String FORM_ID_ATTRIBUTE = "_id";
   public static final String XMLHTTP_REQUEST = "XMLHttpRequest";
   public static final String X_REQUESTED_WITH = "X-Requested-With";
