+++ knotx-core/knotx-template-engine/src/main/java/com/cognifide/knotx/engine/impl/TemplateEngine.java
@@ -17,8 +17,8 @@
  */
 package com.cognifide.knotx.engine.impl;
 
-import com.cognifide.knotx.api.CustomHandlebarsHelper;
-import com.cognifide.knotx.api.TemplateEngineRequest;
+import com.cognifide.knotx.handlebars.CustomHandlebarsHelper;
+import com.cognifide.knotx.dataobjects.TemplateEngineRequest;
 import com.cognifide.knotx.engine.TemplateEngineConfiguration;
 import com.cognifide.knotx.engine.TemplateEngineConsts;
 import com.cognifide.knotx.engine.parser.HtmlFragment;
