+++ knotx-core/knotx-template-engine/src/main/java/com/cognifide/knotx/engine/impl/TemplateSnippetProcessor.java
@@ -17,7 +17,7 @@
  */
 package com.cognifide.knotx.engine.impl;
 
-import com.cognifide.knotx.api.TemplateEngineRequest;
+import com.cognifide.knotx.dataobjects.TemplateEngineRequest;
 import com.cognifide.knotx.engine.TemplateEngineConfiguration;
 import com.cognifide.knotx.engine.parser.HtmlFragment;
 import com.cognifide.knotx.engine.service.ServiceEngine;
