+++ knotx-core/knotx-template-engine/src/main/java/com/cognifide/knotx/engine/placeholders/PlaceholderSubstitutor.java
@@ -17,7 +17,7 @@
  */
 package com.cognifide.knotx.engine.placeholders;
 
-import com.cognifide.knotx.api.TemplateEngineRequest;
+import com.cognifide.knotx.dataobjects.TemplateEngineRequest;
 
 public interface PlaceholderSubstitutor {
