+++ knotx-core/knotx-template-engine/src/main/java/com/cognifide/knotx/engine/service/ServiceAttributeUtil.java
@@ -18,7 +18,7 @@
 package com.cognifide.knotx.engine.service;
 
 
-import com.cognifide.knotx.api.ServiceCallMethod;
+import com.cognifide.knotx.dataobjects.ServiceCallMethod;
 
 import org.apache.commons.lang3.StringUtils;
