+++ knotx-core/knotx-template-engine/src/main/java/com/cognifide/knotx/engine/service/ServiceEntry.java
@@ -18,8 +18,8 @@
 package com.cognifide.knotx.engine.service;
 
 
-import com.cognifide.knotx.api.ServiceCallMethod;
-import com.cognifide.knotx.api.TemplateEngineRequest;
+import com.cognifide.knotx.dataobjects.ServiceCallMethod;
+import com.cognifide.knotx.dataobjects.TemplateEngineRequest;
 import com.cognifide.knotx.engine.TemplateEngineConfiguration;
 import com.cognifide.knotx.engine.TemplateEngineConsts;
 import com.cognifide.knotx.engine.parser.HtmlFragment;
