+++ knotx-core/knotx-template-engine/src/test/java/com/cognifide/knotx/engine/service/ServiceEngineTest.java
@@ -19,7 +19,7 @@ package com.cognifide.knotx.engine.service;
 
 import com.google.common.collect.Lists;
 
-import com.cognifide.knotx.api.TemplateEngineRequest;
+import com.cognifide.knotx.dataobjects.TemplateEngineRequest;
 import com.cognifide.knotx.engine.AbstractKnotxConfigurationTest;
 import com.cognifide.knotx.engine.TemplateEngineConfiguration;
