+++ knotx-core/knotx-template-engine/src/test/java/com/cognifide/knotx/engine/service/ServiceEntryTest.java
@@ -18,7 +18,7 @@
 package com.cognifide.knotx.engine.service;
 
 
-import com.cognifide.knotx.api.TemplateEngineRequest;
+import com.cognifide.knotx.dataobjects.TemplateEngineRequest;
 import com.cognifide.knotx.engine.AbstractKnotxConfigurationTest;
 import com.cognifide.knotx.engine.TemplateEngineConfiguration;
 import com.cognifide.knotx.engine.parser.HtmlFragment;
